refactor(settings): extract About tab — closes #332 phase 1 (slice 6, final)
Final mechanical slice of the settings monolith decomposition.
Moves the About tab — app metadata loaders, manual update probe,
the menu-driven \`updater:result\` listener, and the static
license/source/credit links — into \`AboutTab.svelte\`.

What moved:
- 5 \`\$state\` declarations (appName, appVersion, tauriVersion,
  updateCheck, updateChecking) plus the \`UpdateCheckResult\`
  tagged-union type.
- \`loadAppMetadata\` + \`onCheckForUpdates\` handlers.
- The \`Events.UpdaterResult\` listener registration + onDestroy
  teardown — moves out of the page's onMount despite the menu's
  \`Check for Updates\` item routing through \`settings:goto-tab\`.
  See AboutTab's header comment for the race-window analysis;
  the probe's HTTPS network floor is well above the listener-
  registration timing margin.
- The full About markup (~150 lines) and its scoped CSS (~125
  lines).

What stayed on the page:
- The \`settings:goto-tab\` listener — still owns the cross-tab
  routing.

## Phase 1 close-out

Cumulative across slices 1–6 (#387, #389, #390, #391, #394, this):

| File                          | Before | After | Δ         |
|-------------------------------|--------|-------|-----------|
| settings/+page.svelte         | 2428   | 503   | -1925 (-79%) |

settings/+page.svelte is now a thin tab-switcher (model-picker
inline, 6 tab components mounted lazily based on \`active\`).
Each tab owns its own state, IPC, and lifecycle.

Phase 2 of #332 (specta evaluation for typed IPC) and the CSS
hoist from #392 stay open as follow-ups.

Refs #332, #392.
